+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Maintenance
+- Improved logging for reminders and the starboard.
+- Reduced the amount of time a starboard message is "cleaned up" for.
 
 ## [4.3.1] - 2020-07-27
 ### Maintenance